rename: util 함수 파일명 변경
